# docs/swagger.yaml
basePath: /
definitions:
...
  model.Permissao:
    enum:
    - ADM
    - TEC
    - SUP
    - INF
    - VOIP
    - IMP
    - CAD
    - USR
    - DEV
    type: string
    x-enum-comments:
      PermADM: Administrador
      PermCAD: Cadastro de usuários
      PermDEV: Desenvolvedor
      PermIMP: Técnico de Impressoras
      PermINF: Técnico de Infraestrutura
      PermSUP: Técnico de Suporte (Help Desk)
      PermTEC: Técnico
      PermUSR: Usuário comum (pode apenas abrir chamados)
      PermVOIP: Técnico de Telefonia
    x-enum-descriptions:
    - Administrador
    - Técnico
    - Técnico de Suporte (Help Desk)
    - Técnico de Infraestrutura
    - Técnico de Telefonia
    - Técnico de Impressoras
    - Cadastro de usuários
    - Usuário comum (pode apenas abrir chamados)
    - Desenvolvedor
    x-enum-varnames:
    - PermADM
    - PermTEC
    - PermSUP
    - PermINF
    - PermVOIP
    - PermIMP
    - PermCAD
    - PermUSR
    - PermDEV
...
  model.Usuario:
    properties:
      atualizadoEm:
        type: string
      avatar:
        type: string
      criadoEm:
        type: string
      email:
        type: string
      id:
        type: string
      login:
        type: string
      nome:
        type: string
      permissao:
        $ref: '#/definitions/model.Permissao'
      status:
        type: boolean
      ultimoLogin:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para gestão de chamados
  title: Gestor de Chamados API
  version: "1.0"
paths:
...
  /usuarios/ativar/{id}:
    patch:
      consumes:
      - application/json
      description: Ativa (reativa) usuário pelo ID (apenas ADM)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Ativa usuário
      tags:
      - usuarios
  /usuarios/atualizar-permissao/{id}:
    patch:
      consumes:
      - application/json
      description: Atualiza permissão do usuário pelo ID (apenas ADM)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Permissão do usuário
        in: body
        name: permissao
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Atualiza permissão do usuário
      tags:
      - usuarios
  /usuarios/atualizar/{id}:
    put:
      consumes:
      - application/json
      description: Atualiza dados do usuário (ADM/TEC/USR conforme regra)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados do usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/model.Usuario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Atualiza usuário
      tags:
      - usuarios
  /usuarios/autorizar/{id}:
    patch:
      consumes:
      - application/json
      description: Autoriza (reativa) usuário pelo ID (apenas ADM)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Autoriza usuário
      tags:
      - usuarios
  /usuarios/buscar-novo/{login}:
    get:
      consumes:
      - application/json
      description: Busca usuário no LDAP e retorna dados (apenas ADM)
      parameters:
      - description: Login do usuário
        in: path
        name: login
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
      summary: Busca usuário novo no LDAP
      tags:
      - usuarios
  /usuarios/buscar-por-id/{id}:
    get:
      consumes:
      - application/json
      description: Retorna usuário pelo ID (apenas ADM)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Usuario'
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Busca usuário por ID
      tags:
      - usuarios
  /usuarios/buscar-tecnicos:
    get:
      consumes:
      - application/json
      description: Retorna lista de técnicos (apenas ADM)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Usuario'
            type: array
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Lista técnicos
      tags:
      - usuarios
  /usuarios/buscar-tudo:
    get:
      consumes:
      - application/json
      description: Retorna lista paginada de usuários (apenas ADM)
      parameters:
      - description: Página
        in: query
        name: pagina
        type: integer
      - description: Limite
        in: query
        name: limite
        type: integer
      - description: Busca
        in: query
        name: busca
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Permissão
        in: query
        name: permissao
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Usuario'
            type: array
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Lista usuários com paginação e filtros
      tags:
      - usuarios
  /usuarios/criar:
    post:
      consumes:
      - application/json
      description: Cria um usuário com os dados fornecidos no corpo da requisição.
      parameters:
      - description: Dados do usuário
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/model.Usuario'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema: {}
        "400":
          description: Bad Request
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Cria um novo usuário
      tags:
      - usuarios
  /usuarios/desativar/{id}:
    delete:
      consumes:
      - application/json
      description: Desativa (soft delete) usuário pelo ID (apenas ADM)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Desativa usuário
      tags:
      - usuarios
  /usuarios/lista-completa:
    get:
      consumes:
      - application/json
      description: Retorna todos os usuários (apenas ADM)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Usuario'
            type: array
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Lista completa de usuários
      tags:
      - usuarios
  /usuarios/valida-usuario:
    get:
      consumes:
      - application/json
      description: Verifica se o usuário está autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
      summary: Valida usuário autenticado
      tags:
      - usuarios
swagger: "2.0"